Use devel version of covr, fix option
Because this package uses C++11 covr needs to set a compiler flags to
turn on coverage. Prior to R-3.4 this option was called CXX1XFLAGS,
however from R-3.4 and on it is called CXX11FLAGS. The CRAN version of
covr only sets CXX1XFLAGS, so C++11 code doesn't get the `--coverage`
flag to turn on coverage tracking, which is why you were not getting
code coverage results.