use boston post 2018 format to parse
with minor change for gender of user. seems to insert files into db but then fails in some later C process. so i think its doing the read_one_line_sf ok but then probably fails later with the next step. also seems to read hubway files, which is probably an issue with the R code on city data to shortname mapping.